BaSP-TG-55: Added check for valid id

+++ src/controllers/employees.js
@@ -1,10 +1,27 @@
+import mongoose from 'mongoose';
 import Employees from '../models/Employees';
+
+const { ObjectId } = mongoose.Types;
+
+const isValidObjectId = (id) => ObjectId.isValid(id) && (String)(new ObjectId(id)) === id;
 
 const getEmployeeById = async (req, res) => {
   try {
     const { id } = req.params;
+    if (!isValidObjectId(id)) {
+      return res.status(400).json({
+        message: `Invalid id: ${id}`,
+        error: true,
+      });
+    }
     const employee = await Employees.findById(id);
-
+    if (!employee) {
+      return res.status(404).json({
+        message: `Employee with id ${id} not found`,
+        data: undefined,
+        error: true,
+      });
+    }
     return res.status(200).json({
       message: `Employee with id:${id} found`,
       data: employee,
@@ -21,12 +38,24 @@ const getEmployeeById = async (req, res) => {
 const editEmployee = async (req, res) => {
   try {
     const { id } = req.params;
+    if (!isValidObjectId(id)) {
+      return res.status(400).json({
+        message: `Invalid id: ${id}`,
+        error: true,
+      });
+    }
     const employee = await Employees.findByIdAndUpdate(
-      { _id: id },
+      id,
       { ...req.body },
       { new: true },
     );
-
+    if (!employee) {
+      return res.status(404).json({
+        message: `Employee with id ${id} not found`,
+        data: undefined,
+        error: true,
+      });
+    }
     return res.status(201).json({
       message: `Employee with id:${id} updated successfully`,
       data: employee,
@@ -43,11 +72,21 @@ const editEmployee = async (req, res) => {
 const deleteEmployee = async (req, res) => {
   try {
     const { id } = req.params;
-    await Employees.findByIdAndDelete(id);
-
-    return res.status(204).json({
-      message: `Employee with id:${id} deleted successfully`,
-    });
+    if (!isValidObjectId(id)) {
+      return res.status(400).json({
+        message: `Invalid id: ${id}`,
+        error: true,
+      });
+    }
+    const employee = await Employees.findByIdAndDelete(id);
+    if (!employee) {
+      return res.status(404).json({
+        message: `Employee with id ${id} not found`,
+        data: undefined,
+        error: true,
+      });
+    }
+    return res.sendStatus(204);
   } catch (err) {
     return res.status(500).json({
       message: `Server Error ${err}`,
